requirements.txt
pytest==7.2.0
playwright==1.46.0
faker==28.4.1
allure-pytest==2.12.0